requirements.txt
flask
torch
torchvision
numpy
pillow
opencv-python-headless
matplotlib
werkzeug
pandas
keras
